feat(llvm): add support for stack data structure
This commit adds support for a stack data structure in LLVM. It includes a new module `globals/stack.rs` that defines the `Stack` struct and its methods. The `Stack` struct has a `stack` field that represents the global array that stores the elements of the stack, a `top` field that represents a pointer to the top of the stack, and a `size` field that represents the maximum size of the stack.

The `Stack` struct has methods to check if the stack is full, store an element in the stack, and get the `Stack` instance. The `Stack` instance is created using the `create` method.

This commit also adds a new module `codegen/llvm/builtins/abort.rs` that defines the `AbortBuiltinFunction` struct and its methods. The `AbortBuiltinFunction` struct has methods to declare and get the `abort` function.

feat: add LLVM code generator

Add LLVM code generator to generate LLVM IR from the AST. The LLVM code generator
is added as a new module in the `codegen` module. The `LLVMCodeGenerator` struct
is defined in the `llvm` module. The `LLVMManager` struct is defined in the
`llvm/manager.rs` file. The `LLVMManager` struct is used to manage the LLVM
context, module, and builder. The `LLVMCodeGenerator` struct implements the
`CodeGenerator` trait. The `generate` method of the `LLVMCodeGenerator` struct
takes an AST as input and generates LLVM IR from it.

The `choose_code_generator` function is removed and replaced with the
`code_generator` function. The `code_generator` function takes a
`CodeGeneratorTarget` enum as input and returns a `Box<dyn CodeGenerator>`. The
`CodeGeneratorTarget` enum has